Partial commit for clean compile (-WPedantic)

+++ PiPedalCommon/src/WifiRegulations.cpp
@@ -4,7 +4,7 @@ using namespace pipedal;
 
 const WifiRule*WifiRegulations::GetRule(int32_t frequencyMhz) const
 {
-    int32_t frequencyKhz = frequencyMhz*1000;
+    uint32_t frequencyKhz = (uint32_t)frequencyMhz*1000;
 
     for (const auto&rule: rules) {
         if (frequencyKhz >= rule.start_freq_khz   && frequencyKhz < rule.end_freq_khz)
